requirements.txt
appnope==0.1.0
bintrees==2.0.7
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
clang==4.0.post1
click==6.7
cycler==0.10.0
Cython==0.27.3
decorator==4.2.1
Django==2.0.2
entrypoints==0.2.3
Flask==0.12.2
gensim==3.1.0
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
image==1.5.17
imread==0.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter-kernel-gateway==2.0.2
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.13.3
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pycnic==0.1.3
pydataset==0.2.0
Pygments==2.2.0
Pympler==0.5
pyparsing==2.2.0
python-dateutil==2.6.1
python-resize-image==1.1.11
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
requests==2.13.0
scikit-learn==0.19.0
scipy==1.0.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.10.0
smart-open==1.5.3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.12.2
widgetsnbextension==3.1.3
word2vec==0.9.2
